pyproject.toml
[project]
name = "think-mcp-server"
version = "0.3.0"
description = "thinkthinking's mcp server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "mcp>=1.3.0",
 "pyyaml>=6.0.2",
 "tiktoken>=0.8.0",
 "pyfiglet>=1.0.2",
 "rich>=13.9.4",
 "openai>=1.3.0",
 "time-machine>=2.13.0",
 "httpx[socks]>=0.28.1",
 "prompt-toolkit>=3.0.43",
 "fastapi>=0.109.0",
 "uvicorn>=0.27.0",
 "think-llm-client>=0.3.7",
 "python-dotenv>=1.0.0",
 "pytest>=8.3.4",
 "pytest-asyncio>=0.25.3",
 "pytest-mock>=3.14.0",
]
[[project.authors]]
name = "zhenjie.ye"
email = "yezhenjie@outlook.de"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
think-mcp-server = "think_mcp_server:main"

[tool.hatch.metadata]
allow-direct-references = true